requirements.txt
celery
Django
requests
python-dotenv
redis
django-celery-beat
psycopg2-binary
beautifulsoup4
lxml
selenium
webdriver_manager
openai